+++ tests/conftest.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import asyncio
+import platform
 import ssl
 import sys
 from collections.abc import Generator, Iterator
@@ -9,6 +10,7 @@ from typing import TYPE_CHECKING, Any
 from unittest.mock import Mock
 
 import pytest
+import sniffio
 import trustme
 from _pytest.fixtures import SubRequest
 from trustme import CA
@@ -17,17 +19,24 @@ if TYPE_CHECKING:
     from blockbuster import BlockBuster
 
 uvloop_marks = []
+uvloop_name = "uvloop"
 try:
-    import uvloop
+    if platform.system() == "Windows":
+        uvloop_name = "winloop"
+        import winloop as uvloop
+    else:
+        import uvloop
 except ImportError:
-    uvloop_marks.append(pytest.mark.skip(reason="uvloop not available"))
+    uvloop_marks.append(pytest.mark.skip(reason=f"{uvloop_name} not available"))
     uvloop = Mock()
 else:
     if hasattr(asyncio.AbstractEventLoop, "shutdown_default_executor") and not hasattr(
         uvloop.loop.Loop, "shutdown_default_executor"
     ):
         uvloop_marks.append(
-            pytest.mark.skip(reason="uvloop is missing shutdown_default_executor()")
+            pytest.mark.skip(
+                reason=f"{uvloop_name} is missing shutdown_default_executor()"
+            )
         )
 
 pytest_plugins = ["pytester"]
@@ -35,9 +44,12 @@ pytest_plugins = ["pytester"]
 asyncio_params = [
     pytest.param(("asyncio", {"debug": True}), id="asyncio"),
     pytest.param(
-        ("asyncio", {"debug": True, "loop_factory": uvloop.new_event_loop}),
+        (
+            "asyncio",
+            {"debug": True, "loop_factory": uvloop.new_event_loop},
+        ),
         marks=uvloop_marks,
-        id="asyncio+uvloop",
+        id=f"asyncio+{uvloop_name}",
     ),
 ]
 if sys.version_info >= (3, 12):
@@ -144,3 +156,16 @@ else:
 
     def no_other_refs() -> list[object]:
         return [sys._getframe(1)]
+
+
+@pytest.fixture
+async def event_loop_implementation_name() -> str | None:
+    try:
+        name = sniffio.current_async_library()
+    except sniffio.AsyncLibraryNotFoundError:
+        return None
+
+    if name == "asyncio":
+        return asyncio.get_running_loop().__module__
+
+    return name